add store to handle payment dialog
