+++ src/store/private/expenses/storeExpenseDeleteExpenseDialog.ts
@@ -1,0 +1,22 @@
+import { create } from "zustand";
+
+interface IStoreExpenseDeleteExpenseDialog {
+  isOpen: boolean;
+  openDialog: () => void;
+  closeDialog: () => void;
+}
+
+const useStoreExpenseDeleteExpenseDialog =
+  create<IStoreExpenseDeleteExpenseDialog>((set) => ({
+    isOpen: false,
+    openDialog: () =>
+      set({
+        isOpen: true,
+      }),
+    closeDialog: () =>
+      set({
+        isOpen: false,
+      }),
+  }));
+
+export default useStoreExpenseDeleteExpenseDialog;
